Refactor: Extract ReadOnlyChatbox component
